pyproject.toml
[project]
name = "corpknow-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "chainlit>=2.10.0",
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg>=3.3.3",
    "pypdf>=6.9.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
]
